ENHANCEMENT: cleaned up OrderItems. Removed some remaining bits of session-based cart functionality/variables that were not needed.
BUGFIX: improved $orderitem->Product() to properly access the product version. Product version is now only applied once an order is placed. Note that in ss2.4 the Versioned class in sapphire still has a bug (debug messages), which can break things if a Product_version doesn't exist. Full fix is to patch the Versioned class to return false when version does not exist.