ENHANCEMENT: improved pricing system to allow price alterations before template usage.
Renamed db field 'Price' to 'BasePrice', and introduced sellingPrice() function and getPrice() function. The get price function just returns the sellingPrice, which is the place to make modifications to the BasePrice before presenting to the user. This could be a temporary product discount, or member/region specific pricing.

CMS field updates also.

relates to #41