Handle missing required values consistently.
There was a discrepency between how has_one and db fields were handled
when required but no matching data was supplied. The former would result
in a filter of (FieldName = 0 OR FieldName = NULL) while the latter
would produce FieldName = null.

The behaviour is now more consistent between the two types and matches what
is documented in the class.

CustomProductTest was failing because there's a default FALSE value set for
Premium that was now being compared with NULL.

It was only a side-effect of the existing code that it would correctly
match the missing value with FALSE. If the default had been TRUE then
the test would have failed.

A longer-term aim would be for MatchObjectFilter to match to default
values when no data is provided but since this is not a documented
feature of MatchObjectFilter, it's not something to implement now.